# Replace FoX with pugixml library