--- a/.gitmodules
+++ b/.gitmodules
@@ -1,3 +1,0 @@
-[submodule "src/xml/fox"]
-	path = src/xml/fox
-	url = https://github.com/mit-crpg/fox.git