--- a/src/constants.F90
+++ b/src/constants.F90
@@ -57,7 +57,6 @@ module constants
   integer, parameter :: MAX_LINE_LEN    = 250
   integer, parameter :: MAX_WORD_LEN    = 150
   integer, parameter :: MAX_FILE_LEN    = 255
-  integer, parameter :: REGION_SPEC_LEN = 1000
 
   ! Maximum number of external source spatial resamples to encounter before an
   ! error is thrown.