--- a/src/pugixml/pugixml_c.cpp
+++ b/src/pugixml/pugixml_c.cpp
@@ -1,0 +1,1 @@
+#include "pugixml_c.hpp"