This pull request replaces the [FoX](https://github.com/andreww/fox) XML library with the [pugixml](http://pugixml.org/) library. As discussed in #796, the performance of FoX on very large XML files is terrible (a 140 MB XML file taking > 20 hours to load), so we need a replacement. After looking at a number of options, I settled on pugixml because it is self-contained, extremely fast, is well-maintained, and has an easy-to-use interface.

pugixml itself is a C++ library that defines a number of classes (`xml_node`, `xml_attribute`, `xml_document`, etc.). In order to use it in OpenMC, what I've done is the following:

1. First, expose key member functions of the C++ classes in C, passing and returning pointers to structs that actually store data for each class. This was very easy because all the pugixml classes have an `internal_object()` method that returns that pointer and a constructor that accepts the struct pointer. The resulting C functions are in `src/pugixml/pugixml_c.hpp`.
2. For each C function, create a Fortran procedure interface using the C interoperability feature of the F2003 standard, namely the `ISO_C_BINDING` intrinsic module. The procedure interfaces are found in `src/pugixml/pugixml_f.F90`. As you'll see there, any arguments or return values of the C functions that are pointers are matched by `type(C_PTR)` in Fortran. This includes strings, since they are pointers to chars.
3. Create Fortran derived types -- `XMLNode`, `XMLAttribute`, `XMLDocument`, and `XMLText` -- that have type-bound procedures mimicking the member functions of the C++ classes. The type-bound procedures in turn call the C functions. These derived types also appear in `src/pugixml/pugixml_f.F90`.
4. Refactor the procedures in `src/xml_interface.F90` to use the new derived types rather than FoX types. I was able to preserve the interface for the most part. There are a few minor changes, e.g. the `get_arraysize_*` functions were replaced with `node_word_count`. A few functions are effectively replaced by type-bound procedures of the new derived types.

Most of the other changes in the source are pretty straightforward. A few points worth mentioning:
- The `pugi::xml_element` and `pugi::xml_text` C++ classes actually have `as_bool()` methods that will check for 1/0, true/false, yes/no. Thus, we don't need to handle this directly in OpenMC anymore and instead can just pass a logical when getting a value. (One caveat: it doesn't seem to allow whitespace- `<run_cmfd> true </run_cmfd>` won't work).
- I was able to get rid of the `REGION_SPEC_LEN` variable that specifies the maximum length of a cell region, `<cell region="..." />`. 
- Along with that, I had to make some changes in our HDF5 interface to handle deferred-length character variables, which are used in the new XML derived types. One change is that if you pass an empty string to write to an HDF5 dataset, it won't be written at all.

Don't be alarmed by the net 14k increase in LOC. FoX alone was 52k lines so this is actually a significant decrease. For the time being, I haven't included pugixml as a git submodule since I don't envision us needing to update it with any frequency. It would be an option though since it's hosted on [github](https://github.com/zeux/pugixml).