@cjosey 2017-03-03
I just did some performance testing with opendeplete.  These results are amazing.
```
38747 lines
Old time =  11.5 seconds
New time =   0.66 seconds

77538 lines
Old time =  31.6 seconds
New time =   0.82 seconds

154811 lines
Old time = 105.4 seconds
New time =   1.22 seconds

308241 lines
New time =   1.92 seconds

1347902 lines (58.5 MB)
New time =   5.51 seconds

6733502 lines (292 MB)
New time =  25.1 seconds
```
It is very clear we switched from an O(n^2) to an O(n) algorithm with this swap.  It's that, or it is disk limited so that algorithm performance isn't even relevant.