@nelsonag 2017-03-05
The code looks good to me. I'd like to build it once before merging. After that, we can welcome in our new XML parser!