@smharper 2017-03-06
Just had a look through this PR.  I like the look of it.

The xml parser is dead.  Long live the xml parser!